Adding a dead-simple API for the Admin.
I'd like to create a version of the front-end that runs using a thick-client like Angular. But, before that, I have to get some sort of API in place so that the Angular app has a way to fetch and update data. This API does _as little as possible_, leaning on the Workflow + Service layers to apply the correct validation (contrasted with the CFML controller layer, which actually transforms data before submitting it). This is because the API assumes that the thick-client will already be submitting the proper format (via JSON payloads). To be clear, validation is _still happening_; just with less ceremony.