Adding an Angular-14 front-end for the Admin.
This is my attempt to add a richer UI to the admin experience of the demo. I'm quite a bit rusty on Angular; but, I have something that I think it "good enough" for the demo. Really, it needs a better design and a better set of design-system widgets. But, I don't want to spend all my time working on this. It was fun... until it wasn't :D Time to move on.